src/data/preprocess.py
"""
Data preprocessing script for the generative AI project.
This module handles data loading, cleaning, and transformation.
"""

import argparse
import yaml
import os
import pandas as pd
from transformers import AutoTokenizer


def load_config(config_path):
    """Load configuration from YAML file."""
    with open(config_path, 'r') as f:
        config = yaml.safe_load(f)
    return config


def load_data(file_path):
    """Load raw data from the specified path."""
    print(f"Loading data from {file_path}")
    
    if file_path.endswith('.csv'):
        return pd.read_csv(file_path)
    elif file_path.endswith('.json'):
        return pd.read_json(file_path)
    elif file_path.endswith('.txt'):
        with open(file_path, 'r', encoding='utf-8') as f:
            return f.readlines()
    else:
        raise ValueError(f"Unsupported file format: {file_path}")


def preprocess_data(data, config):
    """Preprocess the data according to the configuration."""
    print("Preprocessing data...")
    
    # Initialize tokenizer
    tokenizer = AutoTokenizer.from_pretrained(config['preprocessing']['tokenizer'])
    
    # Process based on data type
    if isinstance(data, pd.DataFrame):
        # Add your DataFrame preprocessing logic here
        pass
    elif isinstance(data, list):
        # Add your text preprocessing logic here
        pass
    
    return data


def split_data(data, config):
    """Split data into train, validation, and test sets."""
    print("Splitting data...")
    
    train_ratio = config['train_split']
    val_ratio = config['validation_split']
    
    # Add your data splitting logic here
    
    return train_data, val_data, test_data


def save_processed_data(train_data, val_data, test_data, output_dir):
    """Save processed data to the specified directory."""
    print(f"Saving processed data to {output_dir}")
    
    os.makedirs(output_dir, exist_ok=True)
    
    # Add your data saving logic here


def main():
    parser = argparse.ArgumentParser(description='Preprocess data for generative AI model')
    parser.add_argument('--config', type=str, default='configs/data_config.yaml',
                        help='Path to the data configuration file')
    args = parser.parse_args()
    
    # Load configuration
    config = load_config(args.config)
    
    # Load raw data
    data = load_data(config['raw_data_path'])
    
    # Preprocess data
    processed_data = preprocess_data(data, config)
    
    # Split data
    train_data, val_data, test_data = split_data(processed_data, config)
    
    # Save processed data
    save_processed_data(train_data, val_data, test_data, config['processed_data_path'])
    
    print("Data preprocessing completed successfully!")


if __name__ == "__main__":
    main()
